Refactoring the client to be a trait
The client itself is now composed of more basic actions against the
Datadog API.
